requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.1
fastapi-cli==0.0.16
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
shellingham==1.5.4
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.37.2
timescaledb==0.0.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
redis[hiredis]>=5.0.0
fastapi-filter[sqlalchemy]>=1.0.0
alembic>=1.13.0

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==33.1.0
aiosqlite==0.20.0
